==> src/notebooks/3 baseline.ipynb <==
<h1>Baseline Model</h1>
<p>Logistic Regression was chosen as the baseline model due to its simplicity, interpretability, and efficiency in handling binary classification tasks.<br> Its linear nature makes it a suitable starting point to establish a baseline performance, allowing for a clear understanding of the dataset's separability.</p>

<h3>Imports</h3>

import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, f1_score
import joblib

<h3>Loading train dataset</h3>

# Load train data
train_dataset = pd.read_csv('kmaml223/train.csv')#.sample(n=1000)

# Features to consider for classification
feature_columns = ['toxic', 'severe_toxic', 'obscene', 'threat', 'insult', 'identity_hate'] 

train_dataset.head()

<h3>Splitting train dataset</h3>

# Splitting train_data into training and testing subsets
train_subset, test_subset = train_test_split(train_dataset, test_size=0.2, random_state=42)

train_subset['comment_text_cleaned'].fillna('', inplace=True)
test_subset['comment_text_cleaned'].fillna('', inplace=True)

X_train = train_subset['comment_text_cleaned']
X_test = test_subset['comment_text_cleaned']

<h3>Creating a Logistic Regression Model for every feature</h3>

<p>Let's store them as well!</p>

models = {}

# Preprocess text data, initialize and fit TfidfVectorizer
tfidf_vectorizer = TfidfVectorizer()
X_train_tfidf = tfidf_vectorizer.fit_transform(X_train)
X_test_tfidf = tfidf_vectorizer.transform(X_test)

# Save the TfidfVectorizer
tfidf_filename = "models/baseline/tfidf_vectorizer.pkl"
joblib.dump(tfidf_vectorizer, tfidf_filename)
print("TfidfVectorizer saved as tfidf_vectorizer.pkl")

for feature in feature_columns:
    print(f"===== Predicting for '{feature}' =====")
    
    y_train = train_subset[feature]  # Select the target feature    
    y_test = test_subset[feature]  # Select the target feature
    
    # Initialize and train Logistic Regression model
    logreg = LogisticRegression(max_iter=1000)
    logreg.fit(X_train_tfidf, y_train)

    # Predict toxicity on test data
    predictions = logreg.predict(X_test_tfidf)

    # Evaluate the model
    accuracy = accuracy_score(y_test, predictions)
    f1 = f1_score(y_test, predictions)
    
    print(f"Accuracy for '{feature}': {accuracy}")
    print(f"F1 Score for '{feature}': {f1}")

    # Store the trained model in the dictionary
    models[feature] = logreg

    # Save the trained model
    model_filename = f"models/baseline/{feature}_model.pkl"
    joblib.dump(logreg, model_filename)
    print(f"Model for '{feature}' saved as {model_filename}")
    print()

<p>Yeey!</p>
