==> src/notebooks/6 make_submission.ipynb <==
<h1>Making a Submission</h1>

<h3>Imports</h3>

from sklearn.metrics import accuracy_score
import pandas as pd
import joblib

<h3>Specifying a Model and an Output File</h3>

MODEL_DIRECTORY = "models/logistic_regression_f1"
SUBMISSION_PATH = "submissions/logistic_regression_f1.csv"

<h3>Making predictions for every feature</h3>

# Features to consider for classification
feature_columns = ['toxic', 'severe_toxic', 'obscene', 'threat', 'insult', 'identity_hate'] 

# Load the test data
test_data = pd.read_csv('kmaml223/test.csv')
test_data['comment_text_cleaned'].fillna('', inplace=True)
X_test = test_data['comment_text_cleaned']

# Load the saved TfidfVectorizer
loaded_tfidf_vectorizer = joblib.load(f"{MODEL_DIRECTORY}/tfidf_vectorizer.pkl")

# Create a dictionary to store predictions
submission = pd.DataFrame()
submission['id'] = test_data['id']

# Iterate through each feature and make predictions
for feature in feature_columns:
    print(f"===== Predicting for '{feature}' =====")
    
    # Load the stored model for the feature
    model_filename = f"{MODEL_DIRECTORY}/{feature}_model.pkl"
    loaded_model = joblib.load(model_filename)
    
    # Transform test data using the pre-fitted TfidfVectorizer
    X_test_tfidf = loaded_tfidf_vectorizer.transform(X_test)
    
    # Make predictions using the loaded model
    predictions = loaded_model.predict(X_test_tfidf)
    
    # Add predictions to the submission DataFrame
    submission[feature] = predictions

# Save predictions to a submission CSV file
submission.to_csv(SUBMISSION_PATH, index=False)
print(f"Predictions saved to '{SUBMISSION_PATH}'")

<h3>Comparing Difference Between Other Submissions</h3>

# List of file paths to compare with the new submission
baseline_files = ['baseline.csv', 'logistic_regression_accuracy.csv']  # Add more file paths as needed
new_submission = pd.read_csv(SUBMISSION_PATH)

for baseline_file in baseline_files:
    baseline = pd.read_csv(f"submissions/{baseline_file}")
    print(f"Comparison with {baseline_file}:")
    for feature in feature_columns:
        accuracy = accuracy_score(baseline[feature], new_submission[feature])
        print(f"Accuracy for '{feature}': {accuracy}")
    print()  # Add a line break between different baseline comparisons
